+++ R/git.R
@@ -444,7 +444,7 @@ git_sitrep <- function() {
 
   # github user ---------------------------------------------------------------
   hd_line("GitHub")
-  auth_token <- gitcreds_token()
+  auth_token <- gh::gh_token()
   have_token <- auth_token != ""
   if (have_token) {
     kv_line("Personal access token", "<discovered>")