+++ R/usethis-defunct.R
@@ -43,6 +43,7 @@ browse_github_pat <- function(...) {
 #' @rdname usethis-defunct
 #' @export
 github_token <- function() {
+  # TODO: revisit when I have a better recommendation
   details <- glue("
     Use {ui_code('gitcreds::gitcreds_get()')}, \\
     {ui_code('gitcreds::gitcreds_set()')}, and \\
@@ -143,9 +144,9 @@ deprecate_warn_auth_token <- function(whos_asking, details = NULL) {
   what <- glue("{whos_asking}(auth_token = )")
 
   auth_token_explanation <- glue("
-    usethis now delegates token lookup to the gitcreds package, which \\
-    retrieves credentials based on a URL.
-    usethis forms this URL based on the current project's Git remotes.
+    usethis now delegates token lookup to the gh package, which retrieves \\
+    credentials based on the targetted host URL.
+    This URL is determined by the current project's Git remotes.
     The {ui_code('auth_token')} argument is ignored and will eventually be \\
     removed.")